# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Velisca Fashion API",
        "description": "Backend API for the Velisca luxury boutique web application. Provides authentication, token management, and password reset functionality.",
        "contact": {
            "name": "Velisca Support",
            "email": "support@velisca.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.velisca.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Password"
                ],
                "summary": "Send password reset link",
                "description": "Sends a password reset email to the user with a reset link.",
                "operationId": "16282844cc9784c83870befa20958a38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset link sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password reset link sent to your email"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error sending reset link",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Password"
                ],
                "summary": "Reset password",
                "description": "Resets the user password using the token received via email.",
                "operationId": "fefe3931bed932268c35d8e4f5cf3677",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password reset successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error resetting password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh access token",
                "description": "Exchange a valid refresh token for a new access token and refresh token pair. The old refresh token is revoked (rotated).",
                "operationId": "d15783dbbd6961c2f8890d218a954094",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshTokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token refreshed successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AuthTokens"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired refresh token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid or expired refresh token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "description": "User resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "1234567890",
                        "nullable": true
                    },
                    "avatar": {
                        "type": "string",
                        "example": "avatars/abc.jpg",
                        "nullable": true
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "AuthTokens": {
                "description": "Access and refresh token pair",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "access_token": {
                        "type": "string",
                        "example": "1|abc123def456..."
                    },
                    "refresh_token": {
                        "type": "string",
                        "example": "a1b2c3d4e5...128chars"
                    },
                    "expires_in": {
                        "type": "integer",
                        "example": 900
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "description": "Standard error response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error description"
                    },
                    "errors": {
                        "type": "object",
                        "example": "{\"email\":[\"The email field is required.\"]}",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "RefreshTokenRequest": {
                "description": "Refresh token payload",
                "properties": {
                    "refresh_token": {
                        "type": "string",
                        "example": "a1b2c3d4e5...128chars"
                    }
                },
                "type": "object"
            },
            "ForgotPasswordRequest": {
                "description": "Forgot password payload",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    }
                },
                "type": "object"
            },
            "ResetPasswordRequest": {
                "description": "Reset password payload",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "token": {
                        "type": "string",
                        "example": "reset-token-from-email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "newSecurePass1"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "newSecurePass1"
                    }
                },
                "type": "object"
            }
        }
    },
...
}